move some stuff to a new parse_common
